Commands file seggregated
